requirements.txt
PyMySQL==0.9.3
python-dotenv==0.10.3
pdfrw==0.4
reportlab==3.5.32